Move atmospheric fog mesh to its own file
