+++ engine/src/star_system.h
@@ -64,26 +64,7 @@ const unsigned int SIM_QUEUE_SIZE = 128;
 bool PendingJumpsEmpty();
 
 
-struct AtmosphericFogMesh
-{
-    std::string meshname;
-    double scale;
-    float  er;
-    float  eg;
-    float  eb;
-    float  ea;
-    float  dr;
-    float  dg;
-    float  db;
-    float  da;
-    double focus;
-    double concavity;
-    int    tail_mode_start;
-    int    tail_mode_end;
-    int    min_alpha;
-    int    max_alpha;
-    AtmosphericFogMesh();
-};
+
 
 struct Statistics
 {
